icerberux/src/main/webapp/app/entities/cerbac-what-action/delete/cerbac-what-action-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { CerbacWhatActionService } from '../service/cerbac-what-action.service';

import { CerbacWhatActionDeleteDialogComponent } from './cerbac-what-action-delete-dialog.component';

describe('Component Tests', () => {
  describe('CerbacWhatAction Management Delete Component', () => {
    let comp: CerbacWhatActionDeleteDialogComponent;
    let fixture: ComponentFixture<CerbacWhatActionDeleteDialogComponent>;
    let service: CerbacWhatActionService;
    let mockActiveModal: NgbActiveModal;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        declarations: [CerbacWhatActionDeleteDialogComponent],
        providers: [NgbActiveModal],
      })
        .overrideTemplate(CerbacWhatActionDeleteDialogComponent, '')
        .compileComponents();
      fixture = TestBed.createComponent(CerbacWhatActionDeleteDialogComponent);
      comp = fixture.componentInstance;
      service = TestBed.inject(CerbacWhatActionService);
      mockActiveModal = TestBed.inject(NgbActiveModal);
    });

    describe('confirmDelete', () => {
      it('Should call delete service on confirmDelete', inject(
        [],
        fakeAsync(() => {
          // GIVEN
          spyOn(service, 'delete').and.returnValue(of({}));

          // WHEN
          comp.confirmDelete(123);
          tick();

          // THEN
          expect(service.delete).toHaveBeenCalledWith(123);
          expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
        })
      ));

      it('Should not call delete service on clear', () => {
        // GIVEN
        spyOn(service, 'delete');

        // WHEN
        comp.cancel();

        // THEN
        expect(service.delete).not.toHaveBeenCalled();
        expect(mockActiveModal.close).not.toHaveBeenCalled();
        expect(mockActiveModal.dismiss).toHaveBeenCalled();
      });
    });
  });
});
